requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
channels==4.2.0
channels_redis==4.2.1
diff-match-patch==20241021
dj-database-url==2.3.0
Django==5.1.4
django-crispy-forms==2.3
django-formtools==2.5.1
django-grappelli==4.0.1
django-import-export==4.3.3
django-jazzmin==3.0.1
django-localflavor==4.0
django-reversion==5.1.0
django-suit==0.2.28
future==1.0.0
gunicorn==23.0.0
httplib2==0.9.2
msgpack==1.1.0
packaging==24.2
pillow==11.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-stdnum==1.20
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
tablib==3.7.0
typing_extensions==4.12.2
whitenoise==6.8.2
xadmin==0.6.1
